+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Fixed
+- [HE-34](https://xwdrblue.atlassian.net/browse/HE-34) — *Remove-filter-behavior*
 - [HE-33](https://xwdrblue.atlassian.net/browse/HE-33) — *Fix alls typo errors*
 - [HE-17](https://xwdrblue.atlassian.net/browse/HE-17) — *Fix for bar account creation*